data_driven/requirements.txt
# NeuralSite Data-Driven 依赖

# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.11.0

# 缓存
redis>=5.0.0

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# 地理空间处理
GDAL>=3.7.0
Shapely>=2.0.0

# HTTP客户端
requests>=2.31.0
httpx>=0.24.0

# 异步文件
aiofiles>=23.0.0

# 区块链
web3>=6.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 其他工具
python-dotenv>=1.0.0
